requirements.txt
streamlit>=1.32.0
snowflake-connector-python>=3.7.0
openai>=1.30.0
scikit-learn>=1.4.0
numpy>=1.26.0
pytest>=8.0.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
